Moved print_bindings into storage.py
