+++ jaxtyping/__init__.py
@@ -30,13 +30,14 @@ from ._array_types import (
     set_array_name_format as set_array_name_format,
 )
 from ._config import config as config
-from ._decorator import jaxtyped as jaxtyped, print_bindings as print_bindings
+from ._decorator import jaxtyped as jaxtyped
 from ._errors import (
     AnnotationError as AnnotationError,
     TypeCheckError as TypeCheckError,
 )
 from ._import_hook import install_import_hook as install_import_hook
 from ._ipython_extension import load_ipython_extension as load_ipython_extension
+from ._storage import print_bindings as print_bindings
 
 
 # Now import Array and ArrayLike
